Move data display components to organized directories
- Created data-display directory with organized subdirectories
- Moved table, progress, slider, calendar, and skeleton components
- Added proper README documentation for each component
- Created consistent barrel exports (index.ts files)
- Added deprecation notices to original component locations
- Updated COMPONENT-MOVEMENT-PLAN.md with progress tracking
