+++ src/components/ui/data-display/index.ts
@@ -1,0 +1,5 @@
+export * from './table';
+export * from './progress';
+export * from './slider';
+export * from './calendar';
+export * from './skeleton';
