+++ src/components/ui/data-display/progress/index.ts
@@ -1,0 +1,1 @@
+export * from './progress';
