+++ src/components/ui/data-display/skeleton/index.ts
@@ -1,0 +1,1 @@
+export * from './skeleton';
