+++ src/components/ui/data-display/table/index.ts
@@ -1,0 +1,1 @@
+export * from './table';
